back/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
supabase==2.3.4
transbank-sdk==1.4.0
gunicorn==21.2.0
Werkzeug==3.0.1
python-dateutil==2.8.2
pytz==2024.1
cryptography==42.0.5
PyJWT==2.8.0
websockets==12.0
aiohttp==3.9.3
asyncio==3.4.3